feed

+++ backend/controllers/post.js
@@ -49,8 +49,18 @@ exports.deletePost = (req, res, next) => {
     Post.findByPk(req.params.id)
     .then(post => {
         if (req.token.userId == post.UserId || req.token.userAdmin) {
-            const filename = post.imageUrl.split('/images/')[1];
-            fs.unlink(`images/${filename}`, () => {
+            if (post.imageUrl != undefined) {
+                const filename = post.imageUrl.split('/images/')[1];
+                fs.unlink(`images/${filename}`, () => {
+                    Post.destroy({
+                        where: {
+                            id: req.params.id
+                        }
+                    })
+                    .then(() => res.status(200).json({ message: 'Message supprimé !'}))
+                    .catch(error => res.status(400).json({ error }));
+                })
+            }else{
                 Post.destroy({
                     where: {
                         id: req.params.id
@@ -58,7 +68,7 @@ exports.deletePost = (req, res, next) => {
                 })
                 .then(() => res.status(200).json({ message: 'Message supprimé !'}))
                 .catch(error => res.status(400).json({ error }));
-            })
+            }
         } else {
             res.status(403).json({ message: "Vous n'avez pas l'autorisation pour supprimé ce post !" })
         }
